pyproject.toml
[tool.poetry]
name = "narvi"
version = "0.1.0"
description = ""
authors = ["approxit <approxit@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.1.2"
psycopg = {extras = ["binary"], version = "^3.2.3"}
djangorestframework = "^3.15.2"

[tool.poetry.group.dev.dependencies]
ruff = "^0.7.1"
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
